literal functions refactor
